// apps/governance.mento.org/src/lib/graphql/subgraph/generated/subgraph.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  BigDecimal: { input: any; output: any };
  BigInt: { input: any; output: any };
  Bytes: { input: any; output: any };
  /**
   * 8 bytes signed integer
   *
   */
  Int8: { input: any; output: any };
  /**
   * A string representation of microseconds UNIX timestamp (16 digits)
   *
   */
  Timestamp: { input: any; output: any };
};

export type AccessControl = {
  __typename?: "AccessControl";
  asAccount: Account;
  id: Scalars["Bytes"]["output"];
  roles: Array<AccessControlRole>;
};
// ...
export type AccessControlRoleMember = {
  __typename?: "AccessControlRoleMember";
  accesscontrolrole: AccessControlRole;
  account: Account;
  id: Scalars["ID"]["output"];
};
// ...
export type Account = {
  __typename?: "Account";
  asAccessControl: Maybe<AccessControl>;
  asGovernor: Maybe<Governor>;
  asLocking: Maybe<Locking>;
  asTimelock: Maybe<Timelock>;
  asVoting: Maybe<VotingContract>;
  delegateChangedEvent: Array<DelegateChanged>;
  delegateChangedFromEvent: Array<DelegateChanged>;
  delegateChangedToEvent: Array<DelegateChanged>;
  delegateVotesChangedEvent: Array<DelegateVotesChanged>;
  delegationFrom: Array<VoteDelegation>;
  delegationTo: Array<VoteDelegation>;
  events: Array<Event>;
  id: Scalars["Bytes"]["output"];
  locks: Array<Lock>;
  membership: Array<AccessControlRoleMember>;
  proposed: Array<Proposal>;
  proposedCalls: Array<ProposalCall>;
  roleGranted: Array<RoleGranted>;
  roleGrantedSender: Array<RoleGranted>;
  roleRevoked: Array<RoleRevoked>;
  roleRevokedSender: Array<RoleRevoked>;
  timelockedCalls: Array<TimelockCall>;
  voteWeigth: Array<VoteWeight>;
  voted: Array<VoteReceipt>;
};
// ...
export type DelegateChanged = Event & {
  __typename?: "DelegateChanged";
  contract: VotingContract;
  delegation: Maybe<VoteDelegation>;
  delegator: Account;
  emitter: Account;
  fromDelegate: Account;
  id: Scalars["ID"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  toDelegate: Account;
  transaction: Transaction;
};
// ...
export type DelegateVotesChanged = Event & {
  __typename?: "DelegateVotesChanged";
  contract: VotingContract;
  delegate: Account;
  emitter: Account;
  id: Scalars["ID"]["output"];
  newValue: Scalars["BigInt"]["output"];
  oldValue: Scalars["BigInt"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
  voteWeight: Maybe<VoteWeight>;
};
// ...
export type Event = {
  emitter: Account;
  id: Scalars["ID"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type Governor = {
  __typename?: "Governor";
  asAccount: Account;
  id: Scalars["Bytes"]["output"];
  mode: Maybe<Scalars["String"]["output"]>;
  proposalCanceled: Array<ProposalCanceled>;
  proposalCreated: Array<ProposalCreated>;
  proposalExecuted: Array<ProposalExecuted>;
  proposalQueued: Array<ProposalQueued>;
  proposals: Array<Proposal>;
  votecast: Array<VoteCast>;
};
// ...
export type Lock = {
  __typename?: "Lock";
  amount: Scalars["BigInt"]["output"];
  cliff: Scalars["Int"]["output"];
  delegate: Account;
  id: Scalars["ID"]["output"];
  lockCreate: Array<LockCreate>;
  lockId: Scalars["BigInt"]["output"];
  locking: Locking;
  owner: Account;
  relock: Array<Relock>;
  relocked: Scalars["Boolean"]["output"];
  replacedBy: Maybe<Lock>;
  replaces: Maybe<Lock>;
  slope: Scalars["Int"]["output"];
  time: Scalars["BigInt"]["output"];
};
// ...
export type Locking = {
  __typename?: "Locking";
  asAccount: Account;
  delegate: Array<Delegate>;
  id: Scalars["Bytes"]["output"];
  lockCreate: Array<LockCreate>;
  locks: Array<Lock>;
  mode: Maybe<Scalars["String"]["output"]>;
  relock: Array<Relock>;
  withdraw: Array<Withdraw>;
};
// ...
export type Proposal = {
  __typename?: "Proposal";
  calls: Array<ProposalCall>;
  canceled: Scalars["Boolean"]["output"];
  description: Scalars["String"]["output"];
  endBlock: Scalars["BigInt"]["output"];
  eta: Maybe<Scalars["BigInt"]["output"]>;
  executed: Scalars["Boolean"]["output"];
  governor: Governor;
  id: Scalars["ID"]["output"];
  metadata: ProposalMetadata;
  proposalCanceled: Array<ProposalCanceled>;
  proposalCreated: Array<ProposalCreated>;
  proposalExecuted: Array<ProposalExecuted>;
  proposalId: Scalars["BigInt"]["output"];
  proposalQueued: Array<ProposalQueued>;
  proposer: Account;
  queued: Scalars["Boolean"]["output"];
  receipts: Array<VoteReceipt>;
  startBlock: Scalars["BigInt"]["output"];
  state: ProposalState;
  supports: Array<ProposalSupport>;
  votecast: Array<VoteCast>;
  votes: ProposalVotes;
};
// ...
export type ProposalCall = {
  __typename?: "ProposalCall";
  calldata: Scalars["Bytes"]["output"];
  id: Scalars["ID"]["output"];
  index: Scalars["Int"]["output"];
  proposal: Proposal;
  signature: Scalars["String"]["output"];
  target: Account;
  value: Scalars["BigDecimal"]["output"];
};
// ...
export type ProposalCanceled = Event & {
  __typename?: "ProposalCanceled";
  emitter: Account;
  governor: Governor;
  id: Scalars["ID"]["output"];
  proposal: Proposal;
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type ProposalCreated = Event & {
  __typename?: "ProposalCreated";
  emitter: Account;
  governor: Governor;
  id: Scalars["ID"]["output"];
  proposal: Proposal;
  proposer: Account;
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type ProposalExecuted = Event & {
  __typename?: "ProposalExecuted";
  emitter: Account;
  governor: Governor;
  id: Scalars["ID"]["output"];
  proposal: Proposal;
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type ProposalMetadata = {
  __typename?: "ProposalMetadata";
  description: Scalars["String"]["output"];
  title: Scalars["String"]["output"];
};

export type ProposalQueued = Event & {
  __typename?: "ProposalQueued";
  emitter: Account;
  eta: Scalars["BigInt"]["output"];
  governor: Governor;
  id: Scalars["ID"]["output"];
  proposal: Proposal;
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export enum ProposalState {
  Active = "Active",
  Canceled = "Canceled",
  Defeated = "Defeated",
  Executed = "Executed",
  Expired = "Expired",
  NoState = "NoState",
  Pending = "Pending",
  Queued = "Queued",
  Succeeded = "Succeeded",
}

export type ProposalSupport = {
  __typename?: "ProposalSupport";
  id: Scalars["ID"]["output"];
  proposal: Proposal;
  support: Scalars["Int"]["output"];
  votes: Array<VoteReceipt>;
  weight: Scalars["BigInt"]["output"];
};
// ...
export type ProposalVotes = {
  __typename?: "ProposalVotes";
  abstain: VoteType;
  against: VoteType;
  for: VoteType;
  total: Scalars["BigInt"]["output"];
};
// ...
export type RoleGranted = Event & {
  __typename?: "RoleGranted";
  account: Account;
  emitter: Account;
  id: Scalars["ID"]["output"];
  role: AccessControlRole;
  sender: Account;
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type RoleRevoked = Event & {
  __typename?: "RoleRevoked";
  account: Account;
  emitter: Account;
  id: Scalars["ID"]["output"];
  role: AccessControlRole;
  sender: Account;
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type Timelock = {
  __typename?: "Timelock";
  asAccount: Account;
  cancelled: Array<TimelockOperationCancelled>;
  executed: Array<TimelockOperationExecuted>;
  id: Scalars["Bytes"]["output"];
  mindelaychange: Array<TimelockMinDelayChange>;
  operations: Array<TimelockOperation>;
  scheduled: Array<TimelockOperationScheduled>;
};
// ...
export type TimelockCall = {
  __typename?: "TimelockCall";
  data: Scalars["Bytes"]["output"];
  executed: Array<TimelockOperationExecuted>;
  id: Scalars["ID"]["output"];
  index: Scalars["BigInt"]["output"];
  operation: TimelockOperation;
  scheduled: Array<TimelockOperationScheduled>;
  target: Account;
  value: Scalars["BigDecimal"]["output"];
};
// ...
export type Transaction = {
  __typename?: "Transaction";
  blockNumber: Scalars["BigInt"]["output"];
  events: Array<Event>;
  id: Scalars["ID"]["output"];
  timestamp: Scalars["BigInt"]["output"];
};
// ...
export type VoteCast = Event & {
  __typename?: "VoteCast";
  emitter: Account;
  governor: Governor;
  id: Scalars["ID"]["output"];
  proposal: Proposal;
  receipt: VoteReceipt;
  support: ProposalSupport;
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
  voter: Account;
};
// ...
export type VoteDelegation = {
  __typename?: "VoteDelegation";
  contract: VotingContract;
  delegateChangedEvent: Array<DelegateChanged>;
  delegatee: Account;
  delegator: Account;
  id: Scalars["ID"]["output"];
};
// ...
export type VoteReceipt = {
  __typename?: "VoteReceipt";
  id: Scalars["ID"]["output"];
  params: Maybe<Scalars["Bytes"]["output"]>;
  proposal: Proposal;
  reason: Scalars["String"]["output"];
  support: ProposalSupport;
  voter: Account;
  weight: Scalars["BigInt"]["output"];
};
// ...
export type VoteType = {
  __typename?: "VoteType";
  participants: Array<Participant>;
  total: Scalars["BigInt"]["output"];
};

export type VoteWeight = {
  __typename?: "VoteWeight";
  account: Maybe<Account>;
  contract: VotingContract;
  delegateVotesChangedEvent: Array<DelegateVotesChanged>;
  id: Scalars["ID"]["output"];
  value: Scalars["BigInt"]["output"];
};
// ...
export type VotingContract = {
  __typename?: "VotingContract";
  asAccount: Account;
  delegateChangedEvent: Array<DelegateChanged>;
  delegateVotesChangedEvent: Array<DelegateVotesChanged>;
  delegation: Array<VoteDelegation>;
  id: Scalars["Bytes"]["output"];
  totalWeight: VoteWeight;
  weight: Array<VoteWeight>;
};